// src/app/models/pacientes.ts
export class IPacientes {
    idPac: number;
    nomPac: string;
    apPatPac: string;
    apMatPac: string;
    FecNacPac: Date;
    sexoPac: string;
    curpPac: string;
    telPac: string;
    correoPac: string;
    tSangrePac:string;
    estCivPac: string;
    ocupacionPac: string;
    notasPac: string;
    archPac: string
}